requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.1
rqdatac==3.1.0
PyYAML==6.0.2
